pyproject.toml
[project]
name = "hand"
version = "1.0.0"
description = "HAND Me the Data: Fast Robot Adaptation via Hand Path Retrieval"
requires-python = ">=3.10"
dependencies = [
    "cotracker",
    "hydra-core>=1.3.2",
    "omegaconf>=2.3.0",
    "torch>=2.0.0",
    "torchvision",
    "opencv-python-headless",
    "matplotlib",
    "wandb",
    "numpy",
    "seaborn",
    "tqdm",
    "blosc",
    "numba",
    "einops",
    "tensorflow>=2.15.0",
    "opencv-python",
]

[dependency-groups]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.8.6",
]

[tool.uv.sources]
cotracker = { path = "co-tracker", editable = true }

[tool.ruff]
target-version = "py310"
line-length = 120
fix = true

[tool.ruff.lint]
select = ["F401"]

[tool.ruff.format]
quote-style = "single"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["hand"]
